[refactor] 변수명 및 Task 엔터티 수정
1. util 패키지 모음

2. Task 엔터티 리팩토링

3. 시큐리티 권한 제한
