--- easytask/src/main/java/com/easytask/easytask/src/task/dto/PostTaskUserMappingReq.java
@@ -1,9 +1,0 @@
-package com.easytask.easytask.src.task.dto;
-
-import lombok.Getter;
-
-@Getter
-public class PostTaskUserMappingReq {
-    private Long task;
-    private Long irumi;
-}
